kaigi-note/backend/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.21.1
SQLAlchemy==2.0.9
psycopg2-binary==2.9.6
alembic==1.10.3
python-dotenv==1.0.0
pyjwt==2.6.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-jose==3.4.0
pydantic[email]==1.10.21
httpx==0.28.1
asyncpg==0.30.0